requirements.txt
# Same as backend/requirements.txt — use from project root for deployment
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.28.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
